requirements.txt
django
qiniu
redis
celery
djangorestframework
Wand
requests
ipython
ipdb

